Fix a cast when returning big negative values in strtol()
There is a similar, more complex issue in strtoll().

+++ src/stdlib.c
@@ -737,7 +737,7 @@ long strtol(char const * str, char ** endptr, int base)
 	ret = _strtoull(str, endptr, base, &neg);
 	if(neg != 0)
 	{
-		if(ret > (unsigned)LONG_MAX + 1)
+		if(ret > (unsigned long)LONG_MAX + 1)
 		{
 			errno = ERANGE;
 			return LONG_MIN;
